win: make sure we keep track of existing subwins when managing a toplevel
At the point when we started to listen for a toplevel's substructure
events, it might already have subwins. So we must query its existing
children and keep track of them.

This also means we don't care about subwins of an unmanaged window
anymore as when it becomes managed we will query that anyways.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>