Add layout manager
The layout manager tracks the overall window layout for each frame,
which can later be used to plan rendering.

The goal is to provides a more principled way of damage tracking. Currently
what we do is just sprinkling `add_damage_from_win` everywhere in the code.
And hope we didn't forget a place. This is error prone and the result
damage is also imprecise. Getting damage right for animations using that
method would be a nightmare.

The layout manager is the first step of solving that problem.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>