Add renderer
The renderer is the orchestrator of all the rendering components, like
the layout manager, the command builder, etc. It takes a layout from the
core, generate commands for it using command builder, potentially
post-process it, and then send the commands off to the backend to be
executed.

Right now this renderer is incomplete, it doesn't support rendering with
damages, or monitor repaint.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>