picom: fix custom animated shaders
the draw timer has to be stopped before calling the draw_callback_impl
function because when there's a custom animated shader it calls the
queue_redraw funtion that calls the schedule_render function that sets
and starts the draw timer that gets immediately stopped otherwise.

this fixes the screen not being redirected back with unredir-if-possible
and a custom animated shader and only the first frame being drawn with
vsync, no-frame-pacing and a custom animated shader.