renderer: implement monitor repaint properly
So far, in old or new backends, monitor repaint only shows you where the
damage regions are. But enabling it actually disables `use-damage`,
meaning if you have a bug with damage, enabling monitor repaint and it
will disappear. And I would have to compare the damage regions with and
without monitor repaint to figure things out.

This commit makes monitor repaint use damage too. This is only possible
with the new v2 renderer and new backend interface.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>