renderer: fix handling of root pixmap changes
I thought to myself, "I will add this later", then proceeded to forget
about this...

The `root_damaged` flag in `struct session` I added in anticipation for
this is actually not enough. Imagine this:

frame 0 -> frame 1 -> frame 2 (root changed) -> frame 3

frame 2 will be rendered with the `root_damaged` flag set, all good.
But `frame 3` wouldn't, and if it has buffer age 2, i.e. it's rendered
on top of the base of `frame 1`, then the result will be wrong because
frame 1 and 3 has different root background. So instead I added a
`root_image_generation` that is incremented every time the root image
changes.

Fixes #1247

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>