win: make sure a mapped window always have a pixmap bound
We used to only bind the pixmap when the window is damaged the first
time, because we assume before that, there is no meanful content, and we
don't render windows that haven't been damaged.

But, if a window is damaged then immediately unmapped, we won't have a
chance to bind the pixmap. If the window needs to be faded out, we will
be in a position where we need to render the window, but don't have its
pixmap.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>