backend: decompose mask into source and target mask
When trying to add source image transformation (e.g. scaling), I
realized it's ambiguous whether the mask should be applied before or
after the transformation. And after further thought, I decided it would
be better to have two separate masks, one for the source, which would be
in the source's coordinate system (i.e. before any transformation); and
one for the target.

And due to how we currently use the mask, I made the source mask an image,
and the target mask a region. Nothing is stopping us from making both a
combination of an image and a region, but we won't be using that
capability anyway.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>