Add possibility for frame opacity to include menubar/toolbar and so on
This adds the possibility to define the frame opacity but also include all colors in the window that match the frame color and make those also transparent. It even supports the possibility to define the tolerance for color difference between the frame and other parts of the window and make those or gradually more opaque or also transparent just like the frame.

NOTE: tested with flat window frame so not sure how it will work with gradient frames and also only added for GLX atm, lets see if it's interesting enough to be included in the core or otherwise will just keep it for myself.

It adds the following config options:

# Enable frame opacity for colors that match the frame
frame-opacity-for-same-colors = true;

# Tolerance for similar colors (0 exact match, 1 all colors, default 0.5)
frame-opacity-for-same-colors-constraint = 0.5;

# Make different colors opaque by a factor of x (default 5)
frame-opacity-for-same-colors-multiplier = 5;

and for them to work you will have to active frame opacity for e.g.
frame-opacity = 0.7;

With these options you can now have blurred transparent frame + menubar + toolbar or on popup like File Open have transparent window background but list of files be opaque and so on.