region: fix region scaling
We claim region_scale_floor returns the largest integer region contained
by the scaling result. But in fact what it's doing is taking the floor of
each rectangle of the region separated, which leaves gaps between
rectangles. This is not what we want.

Just always take the ceil instead, hopefully the difference is small
enough to be unnoticeable.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>