core: try to recover when we become out-of-sync with the X server
Due to unknown reasons (X server bug?), we observed windows being
destroyed while we had the X server grabbed. In theory, grabbing the
server prevents it from processing requests from any clients that are
not us, which means they shouldn't be able to destroy windows.

Apparently something weird is going on and that's not actually true. So
far we have only saw windows being destroyed, so that's the case we are
fixing in this commit. Whenever we encounter a non-existent window in
`wm_complete_import`, we pause window importing and again try to process
all events the X server has sent us. Which should bring us back into
sync.

Other state changes might also be happening while we have the X server
grabbed. None have been observed so far, so we don't know yet.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>