x: add support for receiving replies as part of the event stream
Why do we need this? This is because, firstly, in the X protocol,
events, errors, and replies are in a single stream to begin with, so it's
not "unnatural" to put them back. And, some replies have to be handled
in the correct "context".

Say, the X window tree look like this:

   A
  / \
 B   C

`C` is the new window, so send a query tree request for it, and you
receive:

> Event: Reparent B to C
> Query tree reply: C has child B

(Remember events, and replies are all in a single stream)

If you handle the reply before the event, which is what will happen
if you do `xcb_query_tree_reply(c, xcb_query_tree(c, ..))`, you will
think window `B` appeared in 2 places.

Also with this change the `x_handle_error` in `ev_handle` is move
into `x_poll_for_event`. It just makes more sense to there.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>