wm: no longer depends on server grabbing for tree consistency
This is part of the effort to remove reliance on the X critical section
for internal consistency. The rational will be explained in later
commits when it is complete.

Instead of using server grabbing to guarantee that our tree is
synchronized with the X server, this commit moves the wm tree to an
"eventually consistent" model, utilizing the newly introduced async X
request infrastructure (See commit: "x: add support for receiving
replies as part of the event stream").

The algorithm is like this: let's look at each tree node individually,
every node has a list of children. If this list consistent with the X
server for each of the tree nodes, the entire tree is consistent.

Let's also define a "partial consistency" concept. A tree is partially
consistent, if for all the nodes in the tree, their lists of chilren are
either consistent, or empty. Meaning some nodes might exist on the X
server, but not replicated on our side. If a node's parent isn't in our
tree, we call that node "orphaned". (Conveniently, a toplevel can never
be orphaned, because the root window is always in the tree.)

To achieve partial consistency, when each node is discovered (either via
a CreateNotify from the server, or via a QueryTree request, you will see
later), we set up an  event mask  on it to  receive further updates, and
then send a QueryTree request for it. The trick is that the reply to the
QueryTree request will  be processed in order with all the other events,
errors and replies. Assuming the tree started out  partially consistent,
it  will still be  partially  consistent  after we processed  the reply.
Because the  reply contains  up-to-date information about  that  node at
that point  in time;  and we have  also processed  all events  that came
before that reply,  so all other nodes are consistent too. Further more,
the tree will  stay partially  consistent, because the event mask we set
up will keep  bringing us new information for the newly discovered node.

This partial consistency will eventually become full consistency after
all of our requests to the X server have completed.

Note this means we might start rendering with a partial window tree. In
practice, that should hardly be noticeable.

(The explanation above glossed over many implementation concerns, some
of those are documentation in code comments.)

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>