wm/tree: fix potential use-after-free in tree changes
If a toplevel is reparented, then destroy, any WM_TREE_CHANGE_CLIENT
events currently queued will reference the freed toplevel node.

Note this doesn't happen when the toplevel is destroyed directly.
Because in that case it will be turned into a zombie, and is guaranteed
to only be freed after the client change event is handled.

With this commit, when a toplevel is killed (i.e. destroyed or
reparented), all currently queued tree changes will be updated to
reference the zombie instead.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>