wm/win: implement EWMH WM_TRANSIENT_FOR extension
In EWMH the meaning of WM_TRANSIENT_FOR is extended, so that when it's
set to the root window or None, the window is considered transient for
all windows in the same window group.

We implement this by using the window group leader as the leader.

Also make win_get_leader_raw robust against invalid leader values.

Fixes #1278

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>