core: make focus update async
This way focus changes are processed in order with all other events, so
we are less likely to run into the issue where the focused window isn't
yet in our window tree.

Known issue: it is still possible for the window to not be in our window
tree. And when that happens, we won't update focus again after the
focused window is imported.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>