core: make creating managed window objects async
This is to solve a race condition that was previously impossible, and
only became possible because the X critical section was removed.

Currently, in `handle_new_windows`, after we created a managed window
object, we check if its map state is already mapped, if so, we call
`win_map_start`. But there is a problem if the window is mapped before
we reached `handle_new_windows`.

There will be an event generated for that map, because managed windows
are toplevels, and we have event mask set up for the root window at the
very beginning, so we won't miss any map events from toplevels. (Except
when picom has just started, and is trying to manage existing windows).

There are two cases, based on whether this map event is handled before
we reach `handle_new_windows`.

1) If the map event is handled after. It itself will trigger another
   `win_map_start` call, which is an error, and we have an assertion for
   that.

But on the other hand, if we don't call `win_map_start` in
`handle_new_windows`:

2) If the map event is handled before. It will do nothing since we
   haven't created a managed window object for that tree node yet. And
   later, since we don't call `win_map_start` in `handle_new_windows`,
   this window will just never be mapped. (This would be the same
   scenario for pre-existing windows when picom just started).

All these problem came from the map event being handled out-of-order
w.r.t. the reply to the GetWindowAttributes request sent by
`win_maybe_allocate`. So we use async requests to make sure they are
handled in order.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>