vblank: alternative invalid present event recovery
I found sometimes, when we receive an invalid PresentCompleteNotify,
keep requesting new present notifys using `last_msc + 1` doesn't really
help us recover. Instead X just keeps sending us bad
PresentCompleteNotifys, and we stuck in a loop. (This is more likely to
happen during screen resolution change. Also this did not seem to be
possible previously, when we were still handling resolution changes
inside the X critical section.)

This commit tries to test out a different approach. The hypothesis is,
maybe doing something to the screen (e.g. render and present a new frame)
helps change the internal state of the X server, thus kicks us out of the
loop. So whenever we get an invalid notify, we fake a vblank event so
rendering can happen. And we keeps doing this until we are (hopefully)
fine.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>