wm/win: don't cache window monitor
We need to update which monitor a window is on when either the window
position/geometry changes, or when monitor configuration changes. We
only did former.

We could fix that by also doing the latter. But there isn't that much
point in caching the monitor anyway, so let's just calculate it when
needed.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>