core: don't schedule vblank callback if we didn't render
We have an optimization where if we didn't render anything for a frame,
(e.g. when the screen is not redirected), we don't set
`ps->backend_busy` to true. This means if there is a damage event right
after, we won't need to wait for the next vblank to update backend_busy
status, and can render immediately.

The problem is, even when `did_render` is false, we still schedule a
vblank callback. If the scenario stated above happens, we would render a
frame before that vblank callback triggers, and after that we will try
to schedule another vblank callback. This creates two callbacks which
would be unnecessary.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>