wm/win: fix handling of windows with multiple window types
ICCCM states a window can have multiple types, which we didn't handle.

Fixing this also includes a slight change to the dbus interface, I hope
that doesn't break too many people.

Also this is pretty unfixable for `wintypes`, and we are moving to
universal rules anyway, so I didn't bother.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>