wm: fix confusion caused by window ID reuse
If a new window reuses the ID of a previously destroyed window, we would
get confused and will not send a query tree request for this new window,
thus misses all of its children. This is because we saved destroyed
window as orphans, and we assume all orphans are either imported or in
the process of being imported, which is not true for destroyed windows.

With this commit we no longer save destroyed windows. We did that
to gracefully handle a window being destroyed with a query tree request
still in-flight. Now we handle it the more obvious way - we mark any
in-flight query trees for omission when a window is destroyed.

(Now, event if we don't save explicitly destroyed windows anymore, some
orphans might get destroyed and have their ID reused without us knowing.
This is because we don't get DestroyNotify for orphans. To catch this
case, we pessimistically re-query orphans that don't have children.)

Even if the window ID is reused after by a newly created window and the
query tree actually succeeds with information from the new window, we
can't use that result anyway, because at the point when the query tree
is answered, we wouldn't have set up the event mask for the new window
yet.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>