wm: fix reparented window not being imported
Before, if we got a reparent event for a previously unknown window, we
will ignore that event because we expect this window will later be found
in a query tree reply. This is not always true.

If a previously unknown window is reparented to a window that is already
fully imported, because we won't send query tree for a fully imported
window, this new will just never be imported.

We will now explicitly initiate the import process in this case. This
also applies to new window reusing the ID of a previously destroyed
window in our orphan tree.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>