wm: avoid ABA problem when importing a window
Consider this scenario:

1. Window created with ID A, which generates a CreateNotify for A.
2. Window with ID A is destroyed.
3. Another window is created, reusing ID A.
4. We receive the CreateNotify for A.

At step 4, we will be querying a different window than what we think we
were querying. It might have a different parent from the one in the
CreateNotify, and we should also ignore the destroy event in (2), as it
wasn't for the window we are querying.

To fix that, when we initiate the import process for a window, we no
longer attach it to its declared parent, instead we orphan it. We also
ignore any destroy event for it while the query tree is in-flight. We
let the query tree reply event decide: 1. what's the window's parent, 2.
whether the window still exists.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>