wm: set StructureNotifyMask on window
Turns out inferring a window's life cycle based on DestroyNotify sent by
its parent is incredibly hard. Especially if the window is at any point
reparented to a window we haven't imported yet, because we won't be
receiving any DestroyNotify for it. And we also can't reliably know if
such reparenting has happened.

So we now set the StructureNotifyMask, which will give us DestroyNotify
sent from the window itself. With this the tracking window life cycle
from the point when the event mask is set becomes easy, and we always
know if a window is destroyed or not.

Setting StructureNotifyMask also means we will get copies of
Circulate/Configure/Map/Unmap/ReparentNotify from the window itself, we
filter this so existing code doesn't have to change. Only for DestroyNotify
we favor the one sent from the window itself.

Also, setting event masks on window MUST be done synchronously. We use
to just send out a ChangeWindowAttributes request and ignore its result.
But it is actually possible, though extremely unlikely, for the
ChangeWindowAttributes to fail but the QueryTree later to succeed, if
the window is destroyed and another created (during the time between we
send these two requests) with the same ID. In that case we would thought
we have set up the event mask, but actually didn't. Which would be
detrimental.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>