wm: children don't get to choose who their parents are
Well, we tried to use query tree replies to attach windows to correct
parents. But that's not good enough. Because the parent can also
actively acquire children via reparent events. The result is windows
fighting each other for control of their parent pointer.

And this is also bad because in we don't know the stacking order of a
window from its query tree reply.

The conclusion is, a window must be fully responsible for updating its
own children list. Because of ABA problem caused by window ID reuse,
it's possible for a window to be attached to a wrong parent temporarily.
This also means receiving a DestroyNotify from a window X about one of its
children with ID Y, doesn't mean the window with ID Y is actually
destroyed, if Y is currently attached to the wrong parent. From that
DestroyNotify, the only thing we can deduce is that X lost a child with
ID Y. Which means events must be handled in two steps:

If we get a DestroyNotify from window X about its child Y, we disconnect
Y from X. If we get a DestroyNotify from window Y for itself, only then
can we really destroy Y.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>