core: don't manage window if it has changed while awaiting X reply
The first thing we do when we get a new toplevel, is sending a
GetWindowAttributes request, and based on its reply, we may allocate a
managed window struct for it. Currently we find which tree node to
attach this managed window struct by its X window ID only, but we should
use the full treeid.

When a new toplevel is added, we queue a creation tree change for it
which contains a pointer to the tree node. If the tree node is destroyed
before the creation tree change is dequeued, we must "cancel" that
creation tree change, because otherwise the pointer in it would be
dangling. This also means there will be no kill tree change for it
either. As a consequence, we have no way to tell anybody about its
zombie node if we create one. So we just don't. If we don't create a
zombie node for a destroyed node, there will be nowhere to put the
managed window struct attached to it if there is one. Which means there
cannot be a one.

Summing it all up - a toplevel cannot have a managed window struct, if
its creation tree change was never dequeued. You can imagine the
creation tree change is handing out a "key" to the tree node, and this
tree node can only have managed window attached to it via this key.

However, if we are getting the tree node via its X window ID only, we
bypass this "key" mechanism, essentially we will find tree node with the
same X ID even if we don't have the key for it yet. This happens if the
window was destroyed, then another window took its ID while we were
waiting for the GetWindowAttributes reply, we shouldn't start managing
it.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>