wm/tree: bump node's gen if it's removed from toplevel
If a node is removed from toplevel then comes back, it will generate a
new creation tree change. Because of the assumption explained in the
previous commit, we can only attach managed window struct to it using
the "key" from the creation tree change. But the node's key hasn't
changed! Since it was toplevel, it will have a previous, older creation
tree change which also handed out this key, which can be used to attach
managed window to it, thus breaks the assumption.

This means every time a node becomes a toplevel, it must have a
different key. We bump its gen number when it's removed from toplevel to
ensure that.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>