wm: make setting event masks async
We care about whether setting the event masks succeeded, we can't
deduced that from the query tree request because of race conditions and
window ID reuse. So we have to check if the request for setting the
event masks itself succeeded or not. But `xcb_request_check` is out of
order w.r.t. to other X events/replies. So essentially we are "looking
into the future" with it, sensing that a window no longer exists before
the usual events and replies we received from X would have told us so.
This breaks assumptions we based many of our assertions on.

For example, in many places when we can't find a window in our tree, we
assume we must have an incomplete tree (i.e. based on the events/replies,
there are windows on X's side we haven't queried). But it's actually
possible that we looked into the future and saw it no longer exists.

We have updated the async request mechanism so it can support requests
that don't expect replies. This commit uses that to make setting event
masks async, so it no longer uses `xcb_request_check` and is no longer
out of order.

Also updated how query tree requests are handled to make it work
together with this change.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>