wm: best effort refreshing leaders even an inconsistent tree
On second thought, it is not very nice to give up as soon as we see the
tree being inconsistent. This could results in flickering, etc.

Instead, still refresh the leaders with best effort. And only set the
flag to retry leader refresh when we actually hit an inconsistent tree
node.

Related #1336

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>