c2: use UINT32_MAX as GetProperty length
Previously we do 2-step property fetch - get length first, then get the
whole property. If the property changes in between, this breaks.

Now we just pass UINT32_MAX as the length, we also stop tracking the max
indices of a property in rules - we just always fetch the whole
property. This makes the code simpler.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>