backend/xrender: be less clever with blur passes
We used to use the target picture as the render target for the last blur
pass, instead of an intermediate picture. This saves a Composite requests.

With this commit, all blur passes are rendered into intermediate
pictures. Then the final blur result is composited into the target. This
is less efficient, but since blur is *essentially* non-functionaly on
xrender (unless you use xf86-video-intel), we can't make it much worse.

This is a preparation for implementing our own accelerated blur on the
xrender backend.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>