c2: refresh window properties after its client window has changed
If a user rule matches against a on-client window property, that property must
be refreshed when a window's client window changed.

No one has reported this bug yet, but this can cause some window to not be
matched correctly.

Changelog: BugFix: Fix sometimes window property changes aren't reflected in
picom.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>