renderer, backend: refactor how shadows are rendered
This commit touches a couple of areas, and unfortunately cannot be broken into
smaller ones since the changes all depends on each other.

To start, a short explanation of how shadows are rendered now. Currently, each
window gets a pre-generated shadow image, this shadow image has shadow-color and
shadow-opacity baked in. Which means changing color or opacity requires
re-generating the shadow image (caveat on shadow-opacity, because it is applied
twice, we can still apply a opacity on the generated shadow image, but that is
wrong and is a bug).

We want to have dynamic shadow colors, which means this has to change.

Firstly, the `opacity` and `dim` parameters for the blit backend operations are
replaced with `tint`, which is a generalization of both parameters. This can be
used to multiply each of the color channels by a separate factor.

Next, we change how shadows are generated. Instead of a color image, we instead
generate a single channel mask image. This is done by mostly removing color,
since what we are doing right now is generating a mask then coloring it, so we
just remove the coloring step. This is fine if the backend support the "blurring
a mask" codepath. But xrender doesn't have that, and is using the "generating a
pixmap -> backend->ops.bind_pixmap" path. This is problematic because
`bind_pixmap` needs a X VisualID, but single channel masks don't have one. This
is why we added a new `new_image_from_pixels` API for uploading CPU pixel bytes
directly into backend images, bypassing `bind_pixmap`. And using it also means
we are no longer limited by the maximum allowed X request size.

Finally, we change how shadows are rendered, by combining the result of previous
two steps - we apply a `tint` to the shadow mask and Voila.

P.S. Implementing tinting in the GL backend is simple, and actually simplifies
the shader code. Only problems is removing `opacity` and `dim` uniforms will
break all existing shaders, and we have kinda committed to a stable shader API.
So, we have to keep them and assign them values that kind of make sense.

Implementing tinting in xrender is more complicated. And it ended up degrading
the performance of color inversion. Color inversion is already slow because it
needs the "Difference" composite operator, which is not hardware accelerated
already, so I think it's a acceptable sacrifice.

Changelog: Internal: Refactor shadow rendering, preparing for dynamic shadow
colors.
Changelog: BugFix: Fix shadow-opacity being applied twice. You may need to
adjust your shadow-opacity settings.
Changelog: BugFix: Properly fix the longstanding "request too big" problem when
uploading shadow images. (Previous workaround: 69b3eee76b6f3ec6be1d2402151d8882279a74a4)

Related: #257

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>