backend/gl: reallocate blur temporary textures less often
We used to reallocate whenever the source image size changes, this was fine
since all we blur is the screen texture. But since we started using blur to
create shadows, this triggers a ton of reallocation because shadows are of all
kinds of sizes.

This commit makes it only reallocate when the source image is too big to fit.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>