backend/gl: tweaking the blur texture reallocation logic
It's a balancing act between number of reallocates (performance) vs
persistent memory usage.

We have 2 kinds of blur contexts: one for the screen image, one for
generating shadows. Screen image rarely changes size, so we want to
reallocate whenever it changes so we use just enough memory; whereas
shadows have all kinds of sizes, so we want to expand the blur textures
in widht and height until everything fits and we can stop reallocating.

These two strategy are in conflict. The second strategy can work for
screen images - until it changes size, say a monitor is removed. In that
case we need some criteria to trigger a reset, so blur textures
allocated for the screen image can shrink, and it can't be too
disruptive to shadow blur textures.

The criteria we chose is the size of the screen image, which means
whenever screen resizes, all textures allocated for shadow blur will be
shrunk too. I think that's a OK trade-off.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>