renderer/command_builder: fix masking of rounded window body for shadows
If `full_shadow = false`, we mask out window body when rendering shadows. For
rounded windows, we remove a corner_radius sized rectangles from the four
corners of the window body. But we forgot to take into account the scaling
factor of the window when removing these corners.

Related: #1389 (partial fix)

Changelog: BugFix: Fix shadow not being correctly masked when scaled.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>