api/backend: add source_mask_scale to backend_blur_args
If a window with rounded corner is transparent, we need to use its mask image
when blurring its background (because it would be silly to decompose rounded
corners into rectangles). If the window is also scaled, then its mask needs to
be scaled too. This was an oversight.

Changelog: Internal: Backend API 2.0 breaking change, backend_blur_args now has
source_mask_scale.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>