script: make `br_once` instructions more robust
`br_once` is intended to only branch the first time a script is
evaluated. So far we've been using the elapsed timer as a proxy for this
(i.e. if elapsed > 0, then don't branch). This is inaccurate and can be
non-deterministic.

Instead, add an explicitly parameter for whether we should branch, and
only set it to true the first time a script is evaluated.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>